modules/cc_setup/python_src/gcs_src/requirements.txt
google-cloud-storage
google-cloud-kms
google-auth
crcmod
